[Rob]Fixed shouldIntakeBall
Previously we moved the isConnected up so that the sensor would not be reconstructed but it would return true if sensor was not working. But then when sensor was reconnected and didn't have good data the sensor was restarted, it also needed to return true..
Further i moved up the !rejectOppColor to return true as it too did not need to reset the color sensor object.